// frontend/src/autogen/proto.d.ts
export class Alert implements IAlert {
  /**
   * Constructs a new Alert.
   * @param [properties] Properties to set
   */
  constructor(properties?: IAlert)

  /** Alert body. */
  public body: string

  /** Alert format. */
  public format: Alert.Format

  /** Alert icon. */
  public icon: string

  /**
   * Creates a new Alert instance using the specified properties.
   * @param [properties] Properties to set
   * @returns Alert instance
   */
  public static create(properties?: IAlert): Alert

  /**
   * Encodes the specified Alert message. Does not implicitly {@link Alert.verify|verify} messages.
   * @param message Alert message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IAlert,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified Alert message, length delimited. Does not implicitly {@link Alert.verify|verify} messages.
   * @param message Alert message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IAlert,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an Alert message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns Alert
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): Alert

  /**
   * Decodes an Alert message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns Alert
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(reader: $protobuf.Reader | Uint8Array): Alert

  /**
   * Verifies an Alert message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an Alert message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns Alert
   */
  public static fromObject(object: { [k: string]: any }): Alert

  /**
   * Creates a plain object from an Alert message. Also converts values to other types if specified.
   * @param message Alert
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: Alert,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this Alert to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for Alert
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}

export namespace Alert {
  /** Format enum. */
  enum Format {
    UNUSED = 0,
    ERROR = 1,
    WARNING = 2,
    INFO = 3,
    SUCCESS = 4,
  }
}
// ...
export interface IArrow {
  /** Arrow data */
  data?: Uint8Array | null

  /** Arrow styler */
  styler?: IStyler | null

  /** Arrow width */
  width?: number | null

  /** Arrow height */
  height?: number | null

  /** Arrow useContainerWidth */
  useContainerWidth?: boolean | null

  /** Arrow id */
  id?: string | null

  /** Arrow columns */
  columns?: string | null

  /** Arrow editingMode */
  editingMode?: Arrow.EditingMode | null

  /** Arrow disabled */
  disabled?: boolean | null

  /** Arrow formId */
  formId?: string | null
}

/** Represents an Arrow. */
export class Arrow implements IArrow {
  /**
   * Constructs a new Arrow.
   * @param [properties] Properties to set
   */
  constructor(properties?: IArrow)

  /** Arrow data. */
  public data: Uint8Array

  /** Arrow styler. */
  public styler?: IStyler | null

  /** Arrow width. */
  public width: number

  /** Arrow height. */
  public height: number

  /** Arrow useContainerWidth. */
  public useContainerWidth: boolean

  /** Arrow id. */
  public id: string

  /** Arrow columns. */
  public columns: string

  /** Arrow editingMode. */
  public editingMode: Arrow.EditingMode

  /** Arrow disabled. */
  public disabled: boolean

  /** Arrow formId. */
  public formId: string

  /**
   * Creates a new Arrow instance using the specified properties.
   * @param [properties] Properties to set
   * @returns Arrow instance
   */
  public static create(properties?: IArrow): Arrow

  /**
   * Encodes the specified Arrow message. Does not implicitly {@link Arrow.verify|verify} messages.
   * @param message Arrow message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IArrow,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified Arrow message, length delimited. Does not implicitly {@link Arrow.verify|verify} messages.
   * @param message Arrow message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IArrow,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an Arrow message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns Arrow
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): Arrow

  /**
   * Decodes an Arrow message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns Arrow
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(reader: $protobuf.Reader | Uint8Array): Arrow

  /**
   * Verifies an Arrow message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an Arrow message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns Arrow
   */
  public static fromObject(object: { [k: string]: any }): Arrow

  /**
   * Creates a plain object from an Arrow message. Also converts values to other types if specified.
   * @param message Arrow
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: Arrow,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this Arrow to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for Arrow
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}

export namespace Arrow {
  /** EditingMode enum. */
  enum EditingMode {
    READ_ONLY = 0,
    FIXED = 1,
    DYNAMIC = 2,
  }
}

/** Properties of a Styler. */
export interface IStyler {
  /** Styler uuid */
  uuid?: string | null

  /** Styler caption */
  caption?: string | null

  /** Styler styles */
  styles?: string | null

  /** Styler displayValues */
  displayValues?: Uint8Array | null
}
// ...
export interface IAudio {
  /** Audio url */
  url?: string | null

  /** Audio startTime */
  startTime?: number | null
}

/** Represents an Audio. */
export class Audio implements IAudio {
  /**
   * Constructs a new Audio.
   * @param [properties] Properties to set
   */
  constructor(properties?: IAudio)

  /** Audio url. */
  public url: string

  /** Audio startTime. */
  public startTime: number

  /**
   * Creates a new Audio instance using the specified properties.
   * @param [properties] Properties to set
   * @returns Audio instance
   */
  public static create(properties?: IAudio): Audio

  /**
   * Encodes the specified Audio message. Does not implicitly {@link Audio.verify|verify} messages.
   * @param message Audio message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IAudio,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified Audio message, length delimited. Does not implicitly {@link Audio.verify|verify} messages.
   * @param message Audio message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IAudio,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an Audio message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns Audio
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): Audio

  /**
   * Decodes an Audio message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns Audio
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(reader: $protobuf.Reader | Uint8Array): Audio

  /**
   * Verifies an Audio message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an Audio message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns Audio
   */
  public static fromObject(object: { [k: string]: any }): Audio

  /**
   * Creates a plain object from an Audio message. Also converts values to other types if specified.
   * @param message Audio
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: Audio,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this Audio to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for Audio
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}
// ...
export namespace Markdown {
  /** Type enum. */
  enum Type {
    UNSPECIFIED = 0,
    NATIVE = 1,
    CAPTION = 2,
    CODE = 3,
    LATEX = 4,
    DIVIDER = 5,
  }
}
// ...
export namespace Metric {
  /** MetricColor enum. */
  enum MetricColor {
    RED = 0,
    GREEN = 1,
    GRAY = 2,
  }

  /** MetricDirection enum. */
  enum MetricDirection {
    DOWN = 0,
    UP = 1,
    NONE = 2,
  }
}
// ...
export namespace NumberInput {
  /** DataType enum. */
  enum DataType {
    INT = 0,
    FLOAT = 1,
  }
}
// ...
export namespace Slider {
  /** DataType enum. */
  enum DataType {
    INT = 0,
    FLOAT = 1,
    DATETIME = 2,
    DATE = 3,
    TIME = 4,
  }

  /** Type enum. */
  enum Type {
    UNSPECIFIED = 0,
    SLIDER = 1,
    SELECT_SLIDER = 2,
  }
}
// ...
export namespace TextInput {
  /** Type enum. */
  enum Type {
    DEFAULT = 0,
    PASSWORD = 1,
  }
}
// ...
export namespace Video {
  /** Type enum. */
  enum Type {
    UNUSED = 0,
    NATIVE = 1,
    YOUTUBE_IFRAME = 2,
  }
}